backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-multipart
requests
python-docx
pypdf2
scikit-learn
reportlab
